rideshare/requirements.txt
openai~=0.27.8
python-dotenv~=1.0.0
colorama~=0.4.6
gradio~=3.37.0
elevenlabs~=0.2.19